認証まで実装

+++ database/migrations/2020_08_10_075219_create_images_table.php
@@ -15,7 +15,7 @@ class CreateImagesTable extends Migration
     {
         Schema::create('images', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger('board_id')->comment('ボードID');
+            $table->integer('board_id')->comment('ボードID');
             $table->string('image_path')->comment('画像パス');
             $table->timestamps();
         });
